requirements.txt
awsebcli==3.20.10
blessed==1.20.0
blinker==1.7.0
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
Flask==3.0.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.26.4
pathspec==0.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
termcolor==1.1.0
urllib3==1.26.18
wcwidth==0.1.9
Werkzeug==3.0.1
zipp==3.17.0
gunicorn==23.0.0
ezdxf
shapely
python-dotenv
boto3
svgpathtools
